% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaprop56.R
\name{betaprop5}
\alias{betaprop5}
\title{En funktion til beregning af beta som det fremgaar i prop 5/6}
\usage{
betaprop5(sigma, mv, pf)
}
\arguments{
\item{sigma}{covariansmatrix}

\item{mv}{vaegte for markedsportefoljen}

\item{pf}{vaegte for portef?ljen man er interesseret i at undersoege}
}
\value{
beta
}
\description{
En funktion til beregning af beta som det fremgaar i prop 5/6
}
\examples{
sigma <- matrix(c(.25,.10,.10,.36),ncol=2)
tangent <- c(0.2857143, 0.7142857)
betaprop5(sigma, tangent, c(1,0))
}
