% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialmodel.R
\name{binomialmodel}
\alias{binomialmodel}
\title{En funktion til beregning af forventet afkast givet en binomialmodel}
\usage{
binomialmodel(prisslut, sandsynlighed = 0.5)
}
\arguments{
\item{prisslut}{En vektor med priser ved sidste tidspunkt}

\item{sandsynlighed}{er for indeværende sat til 0.5 (fremtidig udvikling)}
}
\value{
En tekststreng med spredning og forventning
}
\description{
En funktion til beregning af forventet afkast givet en binomialmodel
}
\examples{
binomialmodel(c(160,131,107.25,87.81,71.89))
}
