% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficientrand.R
\name{efficientrand}
\alias{efficientrand}
\title{En funktion til beregning af efficient rand}
\usage{
efficientrand(mu, sigma)
}
\arguments{
\item{mu}{En vektor med forventede afkastrater}

\item{sigma}{covariansmatrix}
}
\value{
En tekststreng der specificerer sammenhæng mellem en porteføljes varians og dens forventede afkast. output skal læses som at mu^i = koefficienten foran den i'te potens af forventet afkast
}
\description{
En funktion til beregning af efficient rand
}
\examples{
mu <- c(0.03,0.05,0.06)
sigma <- matrix(c(0.04,0.01,0.01,0.01,0.09,0.01,0.01,0.01,0.16), ncol =3)
efficientrand(mu,sigma)
}
