% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbinomtree.R
\name{genbinomtree}
\alias{genbinomtree}
\title{Beregner binomialtræ for aktiepriser}
\usage{
genbinomtree(S_0, alpha, sigma, deltat, N)
}
\arguments{
\item{S_0}{Kurs til t=0}

\item{alpha}{alphaparameter til st.binomialmodel}

\item{sigma}{volparameter i st.binomialmodel}

\item{deltat}{tidsskridt i modellen}

\item{N}{antal tidsskridt i modellen}
}
\value{
matrix med binomialtræet
}
\description{
Beregner binomialtræ for aktiepriser
}
\examples{
sigma <- 0.2
alpha <- 0.07
deltat <- 0.25
S_0 <- 100
N <- 4
genbinomtree(S_0,alpha,sigma,deltat,N)
}
