% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optiontree.R
\name{optiontree}
\alias{optiontree}
\title{Beregner prisen på en option givet dens payoffstruktur}
\usage{
optiontree(terminal, r, q, pathdependent = F, periods = NULL)
}
\arguments{
\item{terminal}{payoff til sidste tidspunkt, oppefra og ned}

\item{r}{vektor af korte risikofrie renter i hver delmodel - angives seperat for hver delmodel, oppefra og ned, startende fra terminal-tidspunkt og gaaende mod venstre i gitteret}

\item{q}{vektor af q-sandsynligheder - angives seperat for hver delmodel, oppefra og ned, startende fra terminal-tidspunkt og gaaende mod venstre i gitteret}

\item{pathdependent}{logisk. skal gitteret udfoldes? Bruges hvis optionens payoff er stiafhaengigt. I dette tilfælde skal der angives terminal-vaerdi for optionen i alle omega'er}

\item{periods}{Antal perioder i modellen. skal kun angives hvis pathdependent = T}
}
\value{
matrix med priser på optionen
}
\description{
Beregner prisen på en option givet dens payoffstruktur
}
\examples{

r <- c(0.075,0.024976,0.05)
q <- c(0.5,0.3408437,0.5)
periods <- 3
terminal <- c(69,14,14,0)
optiontree(terminal,r,q,TRUE,3)
##Det er instruktivt at se at dette er det samme som
terminal <- c(69,14,0)
optiontree(terminal,r,q)
}
