% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfweights.R
\name{pfweights}
\alias{pfweights}
\title{Beregner porteføljevægte for minimum varians-portefølje givet en specificeret afkastrate}
\usage{
pfweights(mu, sigma, mu_p)
}
\arguments{
\item{mu}{En vektor med forventede afkastrater}

\item{sigma}{covariansmatrix}

\item{mu_p}{Krævet afkastrate}
}
\value{
En vektor af porteføljevægte
}
\description{
Beregner porteføljevægte for minimum varians-portefølje givet en specificeret afkastrate
}
\examples{
mu <- c(0.03,0.05,0.06)
sigma <- matrix(c(0.04,0.01,0.01,0.01,0.09,0.01,0.01,0.01,0.16), ncol =3)
mu_p <-0.08
}
