% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qprop.R
\name{qprop}
\alias{qprop}
\title{Beregner q-sandsynligheder}
\usage{
qprop(S0, Sup, Sdown, r = 0)
}
\arguments{
\item{S0}{Kurs til t=0}

\item{Sup}{Kurs i up tilstand t=1}

\item{Sdown}{Kurs i down tilstand t=1}

\item{r}{risikofrirente}
}
\value{
tibble med u,d,q
}
\description{
Beregner q-sandsynligheder
}
\examples{
qprop(100,120,80,0)
}
