% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangentpf.R
\name{tangentpf}
\alias{tangentpf}
\title{En funktion til beregning af vægtene for tangentporteføljen}
\usage{
tangentpf(mu, sigma, riskfree)
}
\arguments{
\item{mu}{En vektor med forventede afkastrater}

\item{sigma}{covariansmatrix}

\item{riskfree}{renten på det risikofri aktiv}
}
\value{
vægtene for tangentporteføljen
}
\description{
En funktion til beregning af vægtene for tangentporteføljen
}
\examples{
mu <- c(0.03,0.05,0.06)
sigma <- matrix(c(0.04,0.01,0.01,0.01,0.09,0.01,0.01,0.01,0.16), ncol =3)
riskfree <- 0.01
tangentpf(mu, sigma,riskfree)

}
