% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covar.R
\name{weighted.covar}
\alias{weighted.covar}
\title{Beregner sandsynlighedsvægtet covarians}
\usage{
weighted.covar(x, y, p)
}
\arguments{
\item{x}{Vektor}

\item{y}{Vektor, hvis denne udelades, sættes y <- x, og man får derved varians}

\item{p}{Vektor af sandsynlighedsvægte}
}
\value{
vægtene for tangentporteføljen
}
\description{
Beregner sandsynlighedsvægtet covarians
}
\examples{
A1 <- c(56,52,51,50,48)
probs <- c(.1,.2,.4,.2,.1)
A1_0 <- 50
weighted.covar(A1/A1_0,p=probs)

}
